Semester-5/Kecerdasan-Buatan/Tugas/sign2speech/requirements.txt
opencv-python
mediapipe
numpy
pandas
scikit-learn
tensorflow
matplotlib
seaborn
gTTS
pyttsx3
playsound==1.3.0
streamlit
streamlit-webrtc
aiortc
av
pyyaml
joblib
Pillow
